requirements.txt
matplotlib==3.5.1
numpy==1.21.2
omegaconf==2.1.1
pickle5==0.0.11
Pillow==9.2.0
pytorch_lightning==1.5.7
regex>2022.1.18
scipy==1.7.3
torch==1.10.0
git+https://github.com/tackgeun/pytorch-meta.git
torchmetrics
torchvision
torch-fidelity==0.3.0
tqdm==4.61.2
configargparse
setuptools==59.5.0
six
easydict
imageio
pymcubes
